Validate requires Google Translate [sh37529]
When making `migrate` work without a Google Translate API Key, I
presumed that `validate` would also work without a API Key.

To me, validation implies that the files are being checked that they all
have the same keys.

However, in reality it means "we send it to Google and check for
errors"